backend/requirements.txt
Django == 3.2
Pillow==9.3.0
djangorestframework==3.14.0
djoser==2.1.0
django-filter==22.1
drf-extra-fields==3.4.0
django-colorfield